+++ anyblok/bloks/io/__init__.py
@@ -33,6 +33,7 @@ class AnyBlokIO(Blok):
 
     @classmethod
     def import_declaration_module(cls):
+        from . import core  # noqa
         cls.declare_io()
         from . import mapping  # noqa
         from . import mixin  # noqa
@@ -42,6 +43,8 @@ class AnyBlokIO(Blok):
 
     @classmethod
     def reload_declaration_module(cls, reload):
+        from . import core
+        reload(core)
         cls.declare_io()
         from . import mapping
         reload(mapping)